requirements.txt
jupyterlab
ipython
dash>=2.8.1
dash_bootstrap_components>=1.3.1
dash_cytoscape>=0.3.0
kafka-python
matplotlib>=3.7.0
numpy>=1.23.5
pandas>=1.5.3
plotly>=5.13.0
PyYAML>=6.0
torch>=2.0.0
tqdm>=4.64.1
